Refactor components
Implement classnames
