+++ src/pages/index.tsx
@@ -1,6 +1,6 @@
 import React, { useContext } from "react"
-import { Layout } from "~/components/layout"
-import Card from "~/components/card/Card"
+import { Layout } from "~/components/Layout"
+import Card from "~/components/Card"
 import { Tool } from "~/model/tool"
 import { SearchContext } from "~/context/SearchContext"
